Extract duplicated code
