+++ src/main/scala/ru/mousecray/endmagic/blocks/portal/BlockMasterStaticPortal.java
@@ -1,46 +1,9 @@
 package ru.mousecray.endmagic.blocks.portal;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.SoundType;
-import net.minecraft.block.material.Material;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.util.EnumBlockRenderType;
-import net.minecraft.util.math.AxisAlignedBB;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IBlockAccess;
-import net.minecraft.world.World;
-import ru.mousecray.endmagic.blocks.BlockWithTile;
 import ru.mousecray.endmagic.tileentity.portal.TileMasterStaticPortal;
 
-public class BlockMasterStaticPortal extends BlockWithTile<TileMasterStaticPortal> {
+public class BlockMasterStaticPortal extends BlockMasterBasePortal<TileMasterStaticPortal> {
     public BlockMasterStaticPortal() {
-        super(Material.ROCK, TileMasterStaticPortal::new);
-        setResistance(8.0F);
-        setHardness(4.0F);
-        setSoundType(SoundType.STONE);
-    }
-
-    @Override
-    public void neighborChanged(IBlockState state, World worldIn, BlockPos pos, Block blockIn, BlockPos fromPos) {
-        if (!worldIn.isRemote)
-            if (worldIn.isBlockPowered(pos))
-                tile(worldIn, pos).openPortal();
-    }
-
-    @Override
-    public EnumBlockRenderType getRenderType(IBlockState state) {
-        return EnumBlockRenderType.MODEL;
-    }
-
-    @Override
-    public boolean isOpaqueCube(IBlockState state) {
-        return false;
-    }
-
-    protected static final AxisAlignedBB DEFAULT_AABB = new AxisAlignedBB(0.0D, 0.0D, 0.0D, 1.0D, 0.9375D, 1.0D);
-
-    @Override
-    public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos) {
-        return DEFAULT_AABB;
+        super(TileMasterStaticPortal::new);
     }
 }
