+++ src/main/scala/ru/mousecray/endmagic/tileentity/portal/TileWithLocation.java
@@ -7,20 +7,20 @@ import ru.mousecray.endmagic.tileentity.EMTileEntity;
 import static ru.mousecray.endmagic.teleport.Location.spawn;
 
 public class TileWithLocation extends EMTileEntity {
-    public Location distination = spawn;
+    public Location destination = spawn;
 
     @Override
     public NBTTagCompound writeToNBT(NBTTagCompound compound) {
-        compound.setTag("distination", distination.toNbt());
+        compound.setTag("destination", destination.toNbt());
         return super.writeToNBT(compound);
     }
 
     @Override
     public void readFromNBT(NBTTagCompound compound) {
-        if (compound.hasKey("distination", 10))
-            distination = Location.fromNbt(compound.getCompoundTag("distination"));
+        if (compound.hasKey("destination", 10))
+            destination = Location.fromNbt(compound.getCompoundTag("destination"));
         else
-            distination = spawn;
+            destination = spawn;
 
         super.readFromNBT(compound);
     }
